move(parse): read_file() to new file
