+++ srcs/parser/parse.c
@@ -1,10 +1,7 @@
-#include <fcntl.h> // open
 #include <stdio.h>
 #include <stdlib.h>
 #include "scene.h"
 #include "libft.h"
-#include "ft_deque.h"
-#include "get_next_line.h"
 #include "result.h"
 #include "debug.h"
 #include "parse.h"
@@ -29,46 +26,6 @@ cy  x,y,z      nx,ny,nz    diameter  height    r,g,b
 	[double]   [-1.0-1.0]  [double]  [double]  [0-255]
 
 */
-
-static void	add_split_line(t_deque *lines, char **line)
-{
-	char			**split_line_with_space;
-	t_deque_node	*node;
-
-	// todo: not allow head space
-	split_line_with_space = ft_split((const char *)*line, ' ');
-	ft_free((void **)line);
-	node = deque_node_new((void *)split_line_with_space);
-	deque_add_back(lines, node);
-}
-
-static t_deque	*read_lines(const int fd)
-{
-	t_deque		*lines;
-	char		*line;
-	t_result	result;
-
-	lines = deque_new();
-	while (true)
-	{
-		line = get_next_line(fd, &result); // todo: if result==FAILURE
-		if (line == NULL)
-			break ;
-		// todo: remove last newline?
-		add_split_line(lines, &line);
-	}
-	return (lines);
-}
-
-static t_deque	*read_file(const char *file_name)
-{
-	int		fd;
-	t_deque	*lines;
-
-	fd = open(file_name, O_RDONLY); // todo: create x_open with error handle
-	lines = read_lines(fd);
-	return (lines);
-}
 
 void	del_lines(void *args)
 {
